refactor: extract ProjectCodeGenerator delegates (687→198 lines)
Extract three package-private delegate classes from ProjectCodeGenerator:
- LauncherTemplate: JavaFX launcher template and file generation (~445 lines)
- GenerationValidator: destination file and Java identifier validation (~143 lines)
- CodeFormatter: NetBeans editor formatting (~104 lines)

Public API (generateCode, generateLauncher) unchanged. All existing tests
pass with zero regressions. New characterization test verifies delegate
extraction preserves generated launcher content and API surface.

Closes #621
